Move observability tests to http-server and validate monitoring runtime behavior
